Modified ZipReader and ZipWriter
I modified ZipReader and ZipWriter to be able to do all the JAXB
handling. For example if I want to get the manifest file with the zip I
simply do:
ZipReader<Manifest> read = new ZipReader<>(zipFile, Manifest.class)
I made use of generics to accomplish this. This removes the need for the
ManifestRunner class since it was doing the same thing. It should be
possible to use these new implementations of the Zip* classes with the Quiz class